# server/docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /users/me:
    get:
      operationId: UsersController_getProfile
      parameters: []
      responses:
        '200':
          description: Returns current user profile
        '401':
          description: Unauthorized
      security:
        - JWT-auth: []
      summary: Get current user profile
      tags: &ref_1
        - users
    patch:
      operationId: UsersController_updateProfile
      parameters: []
      requestBody:
        required: true
        description: User profile data with optional avatar
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Nguyễn Văn A
                email:
                  type: string
                  example: user@example.com
                avatar:
                  type: string
                  format: binary
      responses:
        '200':
          description: Profile updated successfully
        '400':
          description: Bad request - validation failed
        '401':
          description: Unauthorized
      security:
        - JWT-auth: []
      summary: Update current user profile
      tags: *ref_1
  /users/me/password:
    patch:
      operationId: UsersController_changePassword
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordDto'
      responses:
        '200':
          description: Password changed successfully
        '400':
          description: Bad request - validation failed or incorrect current password
        '401':
          description: Unauthorized
      security:
        - JWT-auth: []
      summary: Change current user password
      tags: *ref_1
  /users:
    get:
      operationId: UsersController_getAllUsers
      parameters: []
      responses:
        '200':
          description: Returns list of all users
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin access required
      security:
        - JWT-auth: []
      summary: Get all users (Admin only)
      tags: *ref_1
    post:
      operationId: UsersController_createUser
      parameters: []
      requestBody:
        required: true
        description: User data with optional avatar
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Nguyễn Văn A
                email:
                  type: string
                  example: user@example.com
                password:
                  type: string
                  example: Password123
                role:
                  type: string
                  enum:
                    - admin
                    - staff
                    - chef
                    - warehouse
                    - customer
                  example: staff
                avatar:
                  type: string
                  format: binary
              required:
                - name
                - email
                - password
                - role
      responses:
        '201':
          description: User created successfully
        '400':
          description: Bad request - validation failed
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin access required
      security:
        - JWT-auth: []
      summary: Create new user (Admin only)
      tags: *ref_1
  /users/{id}:
    patch:
      operationId: UsersController_updateUser
      parameters:
        - name: id
          required: true
          in: path
          description: User ID
          schema:
            type: string
      requestBody:
        required: true
        description: Updated user data with optional avatar
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Nguyễn Văn A
                email:
                  type: string
                  example: user@example.com
                role:
                  type: string
                  enum:
                    - admin
                    - staff
                    - chef
                    - warehouse
                    - customer
                  example: staff
                avatar:
                  type: string
                  format: binary
      responses:
        '200':
          description: User updated successfully
        '400':
          description: Bad request - validation failed
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin access required or cannot edit own role
        '404':
          description: User not found
      security:
        - JWT-auth: []
      summary: Update user (Admin only)
      tags: *ref_1
    delete:
      operationId: UsersController_deleteUser
      parameters:
        - name: id
          required: true
          in: path
          description: User ID
          schema:
            type: string
      responses:
        '200':
          description: User deleted successfully
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin access required or cannot delete own account
        '404':
          description: User not found
      security:
        - JWT-auth: []
      summary: Delete user (Admin only)
      tags: *ref_1
...
info:
  title: Restaurant Management API
  description: |2-

        # Restaurant Management System API Documentation
        
        Hệ thống quản lý nhà hàng toàn diện với các tính năng đầy đủ cho việc vận hành nhà hàng hiện đại.
        
        ## Tính năng chính
        - **🔐 Xác thực & Phân quyền**: JWT-based authentication với phân quyền theo vai trò
        - **👥 Quản lý người dùng**: Admin, Staff, Kitchen, Waiter, Warehouse roles
        - **🍽️ Quản lý thực đơn**: Categories, dishes, menus với theo dõi nguyên liệu
        - **📋 Quản lý đơn hàng**: Đặt bàn qua QR, workflow bếp, theo dõi trạng thái
        - **📦 Quản lý kho**: Theo dõi nguyên liệu, quản lý lô hàng, nhập/xuất kho
        - **💳 Thanh toán**: Tích hợp VNPay cho thanh toán online
        - **🪑 Quản lý bàn**: Tạo QR code, theo dõi trạng thái bàn
        - **📊 Báo cáo**: Báo cáo tài chính, báo cáo kho
        
        ## Xác thực API
        Hầu hết các endpoint yêu cầu xác thực qua JWT token. Thêm token vào header Authorization:
        `Authorization: Bearer <your-jwt-token>`
        
        ## Phân quyền người dùng
        - **ADMIN**: Toàn quyền hệ thống
        - **STAFF**: Các thao tác nhân viên phục vụ
        - **KITCHEN**: Các thao tác đặc thù bếp
        - **WAITER**: Quản lý bàn và đơn hàng
        - **WAREHOUSE**: Quản lý kho
        
        ## HTTP Status Codes
        API trả về các mã trạng thái HTTP chuẩn:
        - **200**: Thành công
        - **201**: Tạo mới thành công
...
servers:
...
components:
  securitySchemes:
    JWT-auth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT
      description: Nhập JWT token (không cần prefix Bearer)
      in: header
  schemas:
    ChangePasswordDto:
      type: object
      properties: {}
...